refactor(r1c): stabilize internal imports in mahamantra_research
Complete canonical import migration within mahamantra_research package:
- Switch all internal cross-references from legacy vibe_core.mahamantra.research.* to canonical vibe_core.mahamantra_research.*
- Fix 14 relative imports (from ..protocols._seed) to absolute imports
- Update docstrings and lazy-load paths in __init__.py
- Migrate 37 files total

Legacy compatibility shim at vibe_core/mahamantra/research/__init__.py remains functional.

Tested:
- Canonical imports: vibe_core.mahamantra_research.* ✓
- Legacy imports: vibe_core.mahamantra.research.* ✓ (via shim)
