+++ vibe_core/mahamantra_research/chemistry.py
@@ -15,7 +15,7 @@ from dataclasses import dataclass
 from typing import Final
 
 # Import THE LAW
-from ..protocols._seed import (
+from vibe_core.mahamantra.protocols._seed import (
     GITA_CHAPTERS,
     HALVES,
     HARE_COUNT,
