+++ vibe_core/mahamantra_research/dna_kmer.py
@@ -36,7 +36,7 @@ __genesis__ = "0xbe5ddc45"  # GenesisByte: parampara % 37 == 0
 import array
 from typing import Final
 
-from ..protocols._seed import QUARTERS, WORDS
+from vibe_core.mahamantra.protocols._seed import QUARTERS, WORDS
 
 # =============================================================================
 # CONSTANTS
