+++ vibe_core/mahamantra_research/lotus_radix_n.py
@@ -73,7 +73,7 @@ __genesis__ = "0xbd0f4898"  # GenesisByte: parampara % 37 == 0
 
 from typing import Any, Final, Generic, Iterator, TypeVar
 
-from ..protocols._seed import WORDS
+from vibe_core.mahamantra.protocols._seed import WORDS
 
 # =============================================================================
 # CONSTANTS
