+++ vibe_core/mahamantra_research/maha_generator.py
@@ -22,7 +22,7 @@ import math
 from typing import Final
 
 # Import THE LAW (7 axioms + primary derivations)
-from ..protocols._seed import (
+from vibe_core.mahamantra.protocols._seed import (
     COSMIC_FRAME,
     FIELD_RESONANCE,
     GITA_CHAPTERS,
